refactor: split cli.py into cli/ package with per-command modules
Break 999-line monolith into focused modules:
- cli/__init__.py: Click group and subcommand registration
- cli/_helpers.py: shared emit_output() utility
- cli/lint.py: setup-eval-lint command (160 lines)
- cli/review.py: setup-eval-review command (150 lines)
- cli/security.py: setup-eval-security command (380 lines)
- cli/skill.py: eval-skill command (170 lines)

No behavior change. All 400 tests pass. Entry point unchanged.
